requirements.txt
pandas
numpy
scikit-learn
flask
requests
python-dotenv
